requirements.txt
Flask
Flask-Cors
Flask-JWT-Extended
bcrypt
python-dotenv
requests